requirements.txt
requests==2.18.4
Django==1.8.7
beautifulsoup4==4.6.0
djangorestframework==3.7.1
